AddItem to cart and Checkout and check constraint

+++ src/controllers/CartControllers.ts
@@ -1,7 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
 import { prisma } from '../utils/prisma';
 import { OrderStatus, Size } from '@prisma/client';
-import { BadRequestError } from '../utils/error';
+import { BadRequestError, InternalServerError } from '../utils/error';
+import { PrismaClientUnknownRequestError } from '@prisma/client/runtime/library';
 
 export async function getUserCartItems(
 	req: Request,
@@ -36,8 +37,6 @@ interface AddToCartRequest {
 	sizeOption: Size;
 }
 
-
-// TODO: don't allow adding more than stockCount
 export async function addItemToCart(
 	req: Request<{}, {}, AddToCartRequest>,
 	res: Response,
@@ -46,6 +45,65 @@ export async function addItemToCart(
 	try {
 		const decodedUser = req.decodedToken!;
 		const { itemId, quantity, colorOption, sizeOption } = req.body;
+
+		if (!itemId || !quantity || !colorOption || !sizeOption) {
+			throw new BadRequestError(
+				'itemId, quantity, colorOption and sizeOption are required'
+			);
+		}
+
+		if (quantity < 1) {
+			throw new BadRequestError('Quantity should be greater than 0');
+		}
+
+		const item = await prisma.item.findUnique({
+			where: {
+				id: itemId,
+			},
+			select: {
+				stockCount: {
+					where: {
+						colorOption,
+						sizeOption,
+					},
+				},
+				colorOptions: true,
+				sizeOptions: true,
+			},
+		});
+
+		if (!item) {
+			throw new BadRequestError('Item not found');
+		}
+
+		if (
+			item.colorOptions.indexOf(colorOption) === -1 ||
+			item.sizeOptions.indexOf(sizeOption) === -1
+		) {
+			throw new BadRequestError(
+				`Item doesn't have this color and size. Chosen color: ${colorOption}, Chosen size: ${sizeOption}. Available colors: [${item.colorOptions.join(
+					', '
+				)}], available sizes: [${item.sizeOptions.join(', ')}]`
+			);
+		}
+
+		/**
+		 * StockCount array should always have one element
+		 * as itemId, colorOption and sizeOption are unique.
+		 * This shoudl also be present as it is present
+		 * in sizeOptions and colorOptions, even if stock is zero with count 0
+		 */
+		if (item.stockCount.length === 0) {
+			throw new InternalServerError(
+				'Item Stock for this color and size not found'
+			);
+		}
+
+		if (item.stockCount[0].count < quantity) {
+			throw new BadRequestError(
+				`Not enough stock. Requested: ${quantity}, Available: ${item.stockCount[0].count}`
+			);
+		}
 
 		const cartItem = await prisma.cartItem.upsert({
 			where: {
@@ -73,12 +131,12 @@ export async function addItemToCart(
 						mediaObjects: true,
 					},
 				},
-			}
+			},
 		});
 
-		return res.status(200).json({ 
-			message: 'Item added to cart',
-			cartItem: cartItem
+		return res.status(200).json({
+			message: 'Item updated to cart',
+			cartItem: cartItem,
 		});
 	} catch (err) {
 		next(err);
@@ -93,6 +151,23 @@ export async function removeItemFromCart(
 	try {
 		const decodedUser = req.decodedToken!;
 		const itemId = parseInt(req.params.itemId);
+
+		if (!itemId) {
+			throw new BadRequestError('itemId is required');
+		}
+
+		const cartItem = await prisma.cartItem.findUnique({
+			where: {
+				itemId_userId: {
+					itemId,
+					userId: decodedUser.user_id,
+				},
+			},
+		});
+
+		if (!cartItem) {
+			throw new BadRequestError('Item not found in cart');
+		}
 
 		await prisma.cartItem.delete({
 			where: {
@@ -155,7 +230,7 @@ export async function checkoutController(
 			throw new BadRequestError('Complete your profile to continue');
 		}
 
-		if(userProfile.cartItems?.length === 0) {
+		if (userProfile.cartItems?.length === 0) {
 			throw new BadRequestError('Cart is empty');
 		}
 
@@ -168,6 +243,49 @@ ${userProfile.address?.state}
 ${userProfile.address?.zipcode}
 `;
 
+		const stockCounts = await prisma.stockCount.findMany({
+			where: {
+				itemId: {
+					in: userProfile.cartItems.map(
+						(cartItem) => cartItem.itemId
+					),
+				},
+			},
+		});
+
+		/**
+		 * This will check if any item is out of stock
+		 */
+		for (const cartItem of userProfile.cartItems) {
+			const stockCount = stockCounts.find(
+				(stockCount) =>
+					stockCount.itemId === cartItem.itemId &&
+					stockCount.colorOption === cartItem.colorOption &&
+					stockCount.sizeOption === cartItem.sizeOption
+			);
+
+			if (!stockCount) {
+				throw new InternalServerError(
+					'Item Stock for this color and size not found'
+				);
+			}
+
+			if (stockCount.count < cartItem.quantity) {
+				throw new BadRequestError(
+					`Not enough stock for itemId: ${cartItem.itemId}. Requested: ${cartItem.quantity}, Available: ${stockCount.count}`
+				);
+			}
+		}
+
+		/**
+		 * Even if above check passes, there is a chance
+		 * that stockCount is updated after the check.
+		 * So, we will update the stockCount in a transaction
+		 * The stockCount table has a check constraint
+		 * that count should not be less than zero.
+		 * So, if the count is less than zero, it will
+		 * revert the transaction.
+		 */
 		const stockUpdationCalls = userProfile.cartItems.map((cartItem) => {
 			return prisma.stockCount.update({
 				where: {
@@ -185,36 +303,52 @@ ${userProfile.address?.zipcode}
 			});
 		});
 
-		const [order] = await prisma.$transaction([
-			prisma.order.create({
-				data: {
-					userId: decodedUser.user_id,
-					status: OrderStatus.processing,
-					address: orderAddress,
-					orderItems: {
-						create: userProfile.cartItems.map((cartItem) => ({
-							itemId: cartItem.itemId,
-							quantity: cartItem.quantity,
-							colorOption: cartItem.colorOption,
-							sizeOption: cartItem.sizeOption,
-							price: cartItem.item.price,
-						})),
-					},
-				},
+		try {
+			const [order, cartItemsDelete, ...stockUpdationResponses] =
+				await prisma.$transaction([
+					prisma.order.create({
+						data: {
+							userId: decodedUser.user_id,
+							status: OrderStatus.processing,
+							address: orderAddress,
+							orderItems: {
+								create: userProfile.cartItems.map(
+									(cartItem) => ({
+										itemId: cartItem.itemId,
+										quantity: cartItem.quantity,
+										colorOption: cartItem.colorOption,
+										sizeOption: cartItem.sizeOption,
+										price: cartItem.item.price,
+									})
+								),
+							},
+						},
 
-				include: {
-					orderItems: true,
-				},
-			}),
-			prisma.cartItem.deleteMany({
-				where: {
-					userId: decodedUser.user_id,
-				},
-			}),
-			...stockUpdationCalls,
-		]);
+						include: {
+							orderItems: true,
+						},
+					}),
+					prisma.cartItem.deleteMany({
+						where: {
+							userId: decodedUser.user_id,
+						},
+					}),
 
-		res.status(200).json({ order });
+					...stockUpdationCalls,
+				]);
+
+			return res.status(200).json({ order });
+		} catch (err: any) {
+			if (
+				err instanceof PrismaClientUnknownRequestError &&
+				err.message.includes('violates check constraint')
+			) {
+				throw new BadRequestError(
+					'Stock not enough. Stock decresed after you added to cart. Please try again'
+				);
+			}
+			throw new InternalServerError('Error While Checking Out');
+		}
 	} catch (err) {
 		next(err);
 	}
